Rework passing of args and auxdata
The way it was working would give an error if you passed in dynamically
allocated strings, because Go won't let you pass garbage-collectable
Go data into C code.

Rework it to do most of the work on the Go side with some C helpers.
It's cleaner and probably faster this way, anyway.

Signed-off-by: Corey Minyard <cminyard@mvista.com>